rename dashboarconfig file
