+++ modules/navigation/DashboardSelection.tsx
@@ -4,7 +4,7 @@ import classNames from 'classnames';
 import Link from 'next/link';
 import { DASHBOARD_TABS } from '../../common/constants/dashboardTabs';
 import { useDelimitatedRoute } from '../../common/utils/router';
-import { useDateConfig } from '../../common/state/dashboardConfig';
+import { useDateConfig } from '../../common/state/dateConfig';
 
 export const DashboardSelection: React.FC = () => {
   const { tab } = useDelimitatedRoute();
